# apps/server/generated/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/study-sets": {
      "post": {
        "operationId": "StudySetController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudySetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Study set created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/StudySetResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a study set",
        "tags": [
          "Study Sets"
        ]
      }
    },
...
    "/study-sets/{setId}/cards": {
      "post": {
        "operationId": "SetFlashcardsController_create",
        "parameters": [
          {
            "name": "setId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFlashcardDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Flashcard created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/FlashcardResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a flashcard in a study set",
        "tags": [
          "Flashcards"
        ]
      },
      "get": {
        "operationId": "SetFlashcardsController_list",
        "parameters": [
          {
            "name": "setId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Flashcards retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/FlashcardResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List flashcards in a study set",
        "tags": [
          "Flashcards"
        ]
      }
    },
    "/study-sets/{setId}/cards/progress": {
      "get": {
        "description": "One row per flashcard including status (NEW / LEARNING / MASTERED), weighted streak, and isStarred. Supports ?starred=true / ?status= filters so clients can narrow the deck server-side (flashcard mode \"only starred\", module page mastery filter) instead of loading everything.",
        "operationId": "SetFlashcardsController_listWithProgress",
        "parameters": [
          {
            "name": "setId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "starred",
            "required": false,
            "in": "query",
            "description": "Only return cards the caller has starred.",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Only return cards in this mastery bucket.",
            "schema": {
              "type": "string",
              "enum": [
                "NEW",
                "LEARNING",
                "MASTERED"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Case-insensitive substring match on term or definition. Empty / whitespace-only strings are ignored.",
            "schema": {
              "example": "photo",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Flashcards with progress retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/FlashcardWithProgressDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List flashcards in a study set folded with the caller's progress",
        "tags": [
          "Flashcards"
        ]
      }
    },
    "/study-sets/{setId}/cards/import/preview": {
      "post": {
        "operationId": "SetFlashcardsController_previewCsv",
        "parameters": [
          {
            "name": "setId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "delimiter",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Preview a CSV before importing — returns headers, first 5 data rows, and a suggested column mapping",
        "tags": [
          "Flashcards"
        ]
      }
    },
    "/study-sets/{setId}/cards/import": {
      "post": {
        "operationId": "SetFlashcardsController_importCsv",
        "parameters": [
          {
            "name": "setId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file",
                  "request"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "request": {
                    "type": "string",
                    "description": "JSON-stringified CsvImportRequestDto"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Import flashcards from a CSV using the supplied column mapping",
        "tags": [
          "Flashcards"
        ]
      }
    },
...
  },
  "info": {
    "title": "Mimir API",
    "description": "Auto-generated OpenAPI specification for the Mimir backend. This file is the source of truth for the frontend type pipeline.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt"
      }
    },
    "schemas": {
...
      "ResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          },
          "data": {
            "type": "object"
          },
          "error": {
            "type": "string"
          },
          "meta": {
            "type": "object"
          }
        },
        "required": [
          "ok",
          "message"
        ]
      },
      "StudySetResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isPrivate": {
            "type": "boolean",
            "description": "true = private, false = public"
          },
          "ownerId": {
            "type": "string",
            "description": "Owner id"
          },
          "ownerName": {
            "type": "string"
          },
          "ownerImg": {
            "type": "string"
          },
          "isOwner": {
            "type": "boolean",
            "example": false
          },
          "isCollected": {
            "type": "boolean",
            "example": true
          },
          "flashcardsCount": {
            "type": "number",
            "example": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "description",
          "isPrivate",
          "ownerId",
          "isOwner",
          "flashcardsCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateStudySetDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Advanced Biology"
          },
          "description": {
            "type": "string",
            "example": "Notes and flashcards for bio exams"
          },
          "language": {
            "type": "string",
            "example": "German"
          },
          "isPrivate": {
            "type": "boolean",
            "example": false,
            "description": "true = private, false = public"
          }
        },
        "required": [
          "title"
        ]
      },
...
      "FlashcardResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "studySetId": {
            "type": "string"
          },
          "term": {
            "type": "string"
          },
          "definition": {
            "type": "string"
          },
          "example": {
            "type": "object",
            "nullable": true
          },
          "phonetic": {
            "type": "object",
            "nullable": true
          },
          "hint": {
            "type": "object",
            "nullable": true
          },
          "imageUrl": {
            "type": "object",
            "nullable": true
          },
          "orderIndex": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "studySetId",
          "term",
          "definition",
          "orderIndex",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateFlashcardDto": {
        "type": "object",
        "properties": {
          "term": {
            "type": "string",
            "example": "Photosynthesis"
          },
          "definition": {
            "type": "string",
            "example": "The process by which green plants convert sunlight into energy."
          },
          "example": {
            "type": "string",
            "example": "Photosynthesis powers most life on Earth."
          },
          "phonetic": {
            "type": "string",
            "example": "/ˌfoʊtəˈsɪnθəsɪs/"
          },
          "hint": {
            "type": "string",
            "example": "Starts with \"photo-\""
          },
          "imageUrl": {
            "type": "string",
            "example": "https://cdn.mimir.app/cards/abc.jpg"
          },
          "orderIndex": {
            "type": "number",
            "example": 3,
            "description": "Position within the set. Auto-assigned if omitted."
          }
        },
        "required": [
          "term",
          "definition"
        ]
      },
      "FlashcardWithProgressDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "term": {
            "type": "string"
          },
          "definition": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "LEARNING",
              "MASTERED"
            ]
          },
          "weightedStreak": {
            "type": "string",
            "example": "1.50",
            "description": "Running weighted-correct streak (0.00 – 5.00+)."
          },
          "isStarred": {
            "type": "boolean",
            "example": false
          },
          "masteredAt": {
            "type": "object",
            "nullable": true,
            "example": "2026-06-30T21:04:11.000Z"
          }
        },
        "required": [
          "id",
          "term",
          "definition",
          "status",
          "weightedStreak",
          "isStarred"
        ]
      },
...
    }
  }
}